+++ hooks/errorHook.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
-php=$(which php)
-dir=$(dirname "$0")
-script="${dir}/../../../occ"
 
-#$php "${dir}/run.php" aria2 start $1 $2 $3
-$php $script aria2 error $1 $2 $3
+php_bin=$(command -v php) || exit 127
+hook_dir=$(CDPATH= cd -- "$(dirname -- "$0")" && pwd) || exit 1
+occ="${hook_dir}/../../../occ"
+
+exec "$php_bin" "$occ" aria2 error "$1" "$2" "$3"